Add files via upload
Heyy Guys....

It is a Small Project with Database (User panel as well as Admin panel).

--------- main.java -------------

Here the main.java contain Flight Booking System Code. So compile the Code in cmd by using "javac main.java". Then use "java main" in cmd. Remember the path of the main.java file where you have store it, then compile and run the file in cmd on that path.

---------- requirement.txt -------------

It contains the requirements to use this source code.